bugfix: make all fields always omitempty (#46)
This generator has a bug/abuse of notation which
hit us pretty hard. If you have a `readOnly` and
`required` field in your OpenAPI spec then submitting
a value in the `create` payload without that field
passes validation (as you'd expect). Unfortunately
the generated serializer skips `omitempty` annotation
on required fields. As a result you when you create
a go object omitting these fields, you end up with things
like `empty string` or `0` which fail validation (even
though you're trying to submit nothing).

This looks like a confusion between a type system
(verify you construct obejcts with the correct things)
and a serializer (send the correct set of things over the
wire)

The correct behaviour seems to be to send omitemptie'd
fields always