Adding DateTimeHelper extension
Prior to this change, the dart generator would not properly handle DateTime arrays. Instead, it would generate code that relies on a method that doesn't exist (DateTime.listFromJson).

This update adds a DateTimeHelper extension to the api_helper.dart file and uses it anywhere there is a list of DateTimes.