14432: Do not add @Valid to Java types in "jaxrs-spec" generator
`@Valid` annotation is only needed for objects, where the cascading
validation is required. For standard Java types and enums it is useless.
Moreover, when those annotations are there for those types - Hibernate
Validator's performance degrades significantly (see:
https://hibernate.atlassian.net/browse/HV-1928 and
https://hibernate.atlassian.net/browse/HV-1933). I've checked other
available generators, and it looks like `spring` generates those
annotations properly. So I took the content of the modified files from
that generator and placed to the `jaxrs-spec` one.