feat: emit swagger spec as embedded go bytearray (#73)
Motivation

In some scenarios you'd like to pass the openapi.yaml
to another library in go, e.g. validator lib that would
check if request matches the schema.

This adds a go package inside `api` folder that
go embeds the yaml and exports a `GetSpec() []bytes`
call that can be used with other libraries

Currently if you want to validate requests in a runtime-go
based server you'd have to include confluentinc/api repo
as a submodule and manually reconcile the version with
the version of the sdk you claim to support. This is error prone