elixir: Update Poison dependency range
The locked version of Poison was released in 2017 and there have been
two major releases since then, 4.0 and 5.0. This allows for *existing*
clients to continue to use 3.0, but opens the range so that newer
versions can be used instead.

Between Poison 3.1.0 and Poison 4.0.0, Poison.Decode.decode/2 was
renamed to Poison.Decode.transform/2, so rather than using Poison.Decode
functions directly, we use a compile-time switch to forward to the old
functions if Poison.Decode.decode/2 is exported and to the new function
if not.

Note that this is *not* the desired end-state; that will be to allow
either Poison or Jason (the latter is much more commonly used in the
Elixir community as of now), but there are definitely some compatibility
issues. (Jason allows for *optional* deserialization into a struct, but
it requires explicit configuration for each such struct.)

I'm not sure of the timing on OpenAPI 7, but I would recommend that
Poison be locked to only `~> 5.0` at that release, and if possible, we
have also added Jason support as a configuration option (and perhaps
even the default).