Add @ts-nocheck hint to typescript files
I (and others -- see #10844) am getting errors on my generated OpenAPI client.

Ideally this shouldn't be necessary, but that would be a more complex fix, for someone more familiar with Typescript and OpenAPI than me.

In the meantime, this implements the "shut up" solution and allows the build to succeed.

I don't know exactly where are all the places where this is necessary (Also I'm only interested in typescript-fetch), so I'm adding it whenever there is a tslint:disable + eslint:disable, which seems to work for me :)