don't treat a missing "type" property as "null" type
Not having a "type" property in a sub-schema is not the same as having
a `"type": null` property.

Using the YAML tag `null` in `"type"` and `"$ref"` properties to express a nullable
type is also not supported by either OAS or JSON Schema. `"type"` MUST be either a
string or an array of strings, `"$ref"` MUST be a string. The YAML tag `null` is
not the same as the string `"null"` and using it to convey the same meaning is not
supported by any of the specifications.

fixes #16466