fixes to kotlin-server generator
In order to get the generated code to compile, I had to make a change to the code generation, and I also found an import that was not in use.

First, the path resources in Paths.kt need to have the @Contextual decorator on the datatypes, in order for the serialization to work properly. If I do not include this decorator, I am unable to compile the generated code.

Second, the import of ApiPrincipal in DefaultApi.kt is not needed when there are no auth methods, and can cause a compile error in some cases.