Add List & ClassVar to always imported typing modules for model_generic
With c6e9a4e1ae87793dda3b251e38b7fe6d636863b9 the __typing class variable got
added to the Python models. However, that commit did not properly add the
`ClassVar` and `List` imports. They are currently not present if
additionalProperties is False, which is independent of the existence of the
`__properties` property.