[csharp] use full qualified name to fix a case where generated code doesn't build
without this change, the code generated for this spec

    openapi: 3.0.3
    info:
     title: test
     version: v0.1
     description: test
    paths:
     /api/path:
      get:
       responses:
        '200':
          description: ok
          content:
           application/json:
            schema:
             $ref: '#/components/schemas/TimeZone'
    components:
      schemas:
       TimeZone:
        type: object
        properties:
          someField:
           type: array
           items:
             type: string

with the default csharp generator (ie with

    java -jar openapi-generator-cli.jar generate --generator-name csharp \
      --input-spec spec.yaml --output out_dir

)
fails to build. ie:

    dotnet build Org.OpenAPITools.sln

fails with an error saying that references to TimeZone are ambiguous,
because they may refer to either our entity, or System.TimeZone.

This commit fixes it by prefixing the entity name with its namespace.

Note that the fix is a bit more subtle than just adding the namespace
every time: consider for instance that other spec:

    openapi: 3.0.3
    info:
     title: test
     version: v0.1
     description: test
    paths:
     /api/path2:
      get:
       responses:
        '200':
          description: ok
          content:
           application/json:
            schema:
             $ref: '#/components/schemas/ObjectIsh'
     /api/path3:
      get:
       responses:
        '200':
          description: ok
          content:
           application/json:
            schema:
             $ref: '#/components/schemas/StringIsh'
    components:
      schemas:
       ObjectIsh:
        type: object
       StringIsh:
        type: string

This does not generates an "ObjectIsh" or a "StringIsh" entity: instead
the generated code directly uses "object" and "string". If we add the
entities namespace everytime, then the generated code would not build.

So this commit manages to add the namespace only when needed (so that
both spec above generate valid C# code)

This fixes #16123