Optimize zero and single-element array initializations
Replace with Collections.singletonList(), Collections.emptyList(), or List.of()
the latter depending on whether content is guaranteed to be constant).
This will save some memory.

In particular, Collections.emptyList() and List.of() with no arguments always
return a shared instance, while Arrays.asList() with no arguments creates a
new object every time it's called.