Improved ClassLoading Conformity
In our use case we are generating documentation at runtime inside of an
application framework that generally does not work well with things like
Class.forName. In an effort to gain consistency in classloading a
ClassLoading utility (ClassLoadingUtil) was added that will prefer the
thread's current classloader before trying other sources. Other sources
can be either a Class or ClassLoader and it will call the appropriate
methods as needed.

This is mainly intended to allow applications that do not have a flat
classpath to offer customization and extension through correct
classloading.