Exclude jackson versions coming from jackson-datatype-threetenbp repository
The versions that are coming from the jackson-datatype-threetenbp are outdated and cause an issue when generating.
There are interopability issues between jackson 2.15.2 and 2.16.0 and above.

The dependency on jackson-threetenbp is causing jacskon 2.15.2 to be needed when generating code. This dependency has not been updated in over a year. Untill it does, it should be excluded from delivering its outdated versions to the classpath.

Fixes https://github.com/OpenAPITools/openapi-generator/issues/18753