When checking for enum default value, compare with the unescaped value
In some representations, the enum value is returned with quotes.
"foo" => "\" + foo + "\".

This bug was fixed in earlier versions but then it broke again when the enum representation was changed.

Ideally, we would check this only against a unescaped value. But to be sure, we do not introduce breaking changes, it is better to check against both.